settings: Extract most of test_settings to test_extra_settings.
